Porting to C# (3/3)
All errors are fixed. Next up, creating a script that can call upon these classes in order to solve for variable.

+++ Assets/Scripts/Non-monobehaviour/UnaryExpression.cs
@@ -15,7 +15,7 @@
     }
 
     override
-    public double Evaluate() {
+    public float Evaluate() {
         // Only runs when isKnown
         return float.Parse(value);
     }
